activate uncovered test
