+++ tests/conftest.py
@@ -8,12 +8,9 @@ import pathlib
 
 import numpy as np
 import pandas as pd
-import plotly.colors as pc
 import pytest
 import skimage
 from PIL import Image
-
-import cytodataframe
 
 
 @pytest.fixture(name="cytotable_CFReT_data_df")
@@ -111,34 +108,6 @@ def fixture_basic_outlier_parquet(
     )
 
     return parquet_path
-
-
-@pytest.fixture(name="generate_show_report_html_output")
-def fixture_generate_show_report_html_output(cytotable_CFReT_data_df: pd.DataFrame):
-    """
-    Used for generating report output for use with other tests.
-    """
-
-    # create outliers dataframe
-    df = cytodataframe.analyze.label_outliers(
-        df=cytotable_CFReT_data_df,
-        include_threshold_scores=True,
-    )
-
-    # show a report
-    df.show_report(
-        report_path=(
-            report_path := pathlib.Path(__file__).parent
-            / "data"
-            / "coSMicQC"
-            / "show_report"
-            / "cosmicqc_example_report.html"
-        ),
-        color_palette=pc.qualitative.Dark24[0:2],
-        auto_open=False,
-    )
-
-    return report_path
 
 
 @pytest.fixture
