+++ tests/test_frame.py
@@ -7,6 +7,8 @@ import pathlib
 import cosmicqc
 import pandas as pd
 import plotly
+import plotly.colors as pc
+import pytest
 from pyarrow import parquet
 
 from cytodataframe.frame import CytoDataFrame
@@ -161,3 +163,31 @@ def test_repr_html(
             "Image_FileName_GOLD",
         ],
     ), "The nuclear speckles images do not contain green outlines."
+
+
+@pytest.mark.generate_report_image
+def fixture_generate_show_report_html_output(cytotable_CFReT_data_df: pd.DataFrame):
+    """
+    Used for generating report output for use with other tests.
+    """
+
+    # create outliers dataframe
+    df = cosmicqc.analyze.label_outliers(
+        df=cytotable_CFReT_data_df,
+        include_threshold_scores=True,
+    )
+
+    # show a report
+    df.show_report(
+        report_path=(
+            report_path := pathlib.Path(__file__).parent
+            / "data"
+            / "coSMicQC"
+            / "show_report"
+            / "cosmicqc_example_report.html"
+        ),
+        color_palette=pc.qualitative.Dark24[0:2],
+        auto_open=False,
+    )
+
+    return report_path
