+++ tests/utils.py
@@ -4,31 +4,13 @@ Utilities for running pytest tests in CytoDataFrame
 
 import base64
 import re
-import subprocess
 from io import BytesIO
-from typing import List, Tuple
+from typing import List
 
 import numpy as np
 from PIL import Image
 
 from cytodataframe import CytoDataFrame
-
-
-def run_cli_command(command: str) -> Tuple[str, str, int]:
-    """
-    Run a CLI command using subprocess and capture the output and return code.
-
-    Args:
-        command (list): The command to run as a list of strings.
-
-    Returns:
-        tuple: (str: stdout, str: stderr, int: returncode)
-    """
-
-    result = subprocess.run(
-        command.split(" "), capture_output=True, text=True, check=False
-    )
-    return result.stdout, result.stderr, result.returncode
 
 
 def cytodataframe_image_display_contains_green_pixels(
